Crud of all Entity

+++ pc/src/main/java/gestion/analytique/pc/repository/CentreRepository.java
@@ -3,5 +3,5 @@ package gestion.analytique.pc.repository;
 import org.springframework.data.repository.CrudRepository;
 import gestion.analytique.pc.model.Centre;
 
-public interface CentreRepository extends CrudRepository<Centre, Long> {
+public interface CentreRepository extends CrudRepository<Centre, Integer> {
 }
